% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{dagri_empty_named_list}
\alias{dagri_empty_named_list}
\title{Create an empty named list}
\usage{
dagri_empty_named_list()
}
\value{
An empty named list.
}
\description{
Utility for initializing empty named list results.
}
\keyword{internal}
