% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{dagri_remove_edge}
\alias{dagri_remove_edge}
\title{Remove an edge from a dagriculture graph}
\usage{
dagri_remove_edge(graph, id)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{id}{Edge ID.}
}
\description{
Remove an edge from a dagriculture graph
}
