% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{groots_add_gate}
\alias{groots_add_gate}
\title{Add a gate to a groots graph}
\usage{
groots_add_gate(graph, edge_id, id = NULL, metadata = list())
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{edge_id}{Edge ID.}

\item{id}{Optional Gate ID.}

\item{metadata}{Gate metadata.}
}
\description{
Add a gate to a groots graph
}
