% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{groots_ancestors}
\alias{groots_ancestors}
\title{Get all ancestors}
\usage{
groots_ancestors(graph, node_id)
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get all ancestors
}
