% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{groots_descendants}
\alias{groots_descendants}
\title{Get all descendants}
\usage{
groots_descendants(graph, node_id)
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get all descendants
}
