% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{groots_downstream}
\alias{groots_downstream}
\title{Get downstream nodes}
\usage{
groots_downstream(graph, node_id)
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get downstream nodes
}
