% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{groots_edge}
\alias{groots_edge}
\title{Get an edge from a groots graph}
\usage{
groots_edge(graph, edge_id)
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{edge_id}{Edge ID.}
}
\description{
Get an edge from a groots graph
}
