% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{groots_gate}
\alias{groots_gate}
\title{Get a gate from a groots graph}
\usage{
groots_gate(graph, id)
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{id}{Gate ID.}
}
\description{
Get a gate from a groots graph
}
