% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{groots_graph}
\alias{groots_graph}
\title{Create an empty groots graph}
\usage{
groots_graph(registry)
}
\arguments{
\item{registry}{A \code{groots_registry} object.}
}
\description{
Create an empty groots graph
}
