% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{groots_node}
\alias{groots_node}
\title{Get a node from a groots graph}
\usage{
groots_node(graph, node_id)
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get a node from a groots graph
}
