% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{groots_plan}
\alias{groots_plan}
\title{Create an execution plan}
\usage{
groots_plan(graph, targets = NULL)
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{targets}{Optional target nodes.}
}
\description{
Create an execution plan
}
