% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{groots_recompute_state}
\alias{groots_recompute_state}
\title{Recompute graph state}
\usage{
groots_recompute_state(graph)
}
\arguments{
\item{graph}{A \code{groots_graph}.}
}
\description{
Recompute graph state
}
