% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{groots_remove_gate}
\alias{groots_remove_gate}
\title{Remove a gate from a groots graph}
\usage{
groots_remove_gate(graph, id)
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{id}{Gate ID.}
}
\description{
Remove a gate from a groots graph
}
