% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{groots_remove_node}
\alias{groots_remove_node}
\title{Remove a node from a groots graph}
\usage{
groots_remove_node(graph, node_id)
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{node_id}{Node ID.}
}
\description{
Remove a node from a groots graph
}
