% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{groots_reopen_gate}
\alias{groots_reopen_gate}
\title{Reopen a gate in a groots graph}
\usage{
groots_reopen_gate(graph, id)
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{id}{Gate ID.}
}
\description{
Reopen a gate in a groots graph
}
