% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{groots_resolve_gate}
\alias{groots_resolve_gate}
\title{Resolve a gate in a groots graph}
\usage{
groots_resolve_gate(graph, id)
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{id}{Gate ID.}
}
\description{
Resolve a gate in a groots graph
}
