% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{groots_topo_order}
\alias{groots_topo_order}
\title{Get topological order}
\usage{
groots_topo_order(graph, subset = NULL)
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{subset}{Optional subset of nodes.}
}
\description{
Get topological order
}
