% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{groots_upstream}
\alias{groots_upstream}
\title{Get upstream nodes}
\usage{
groots_upstream(graph, node_id)
}
\arguments{
\item{graph}{A \code{groots_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get upstream nodes
}
