% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{dagri_blocked}
\alias{dagri_blocked}
\title{Get blocked nodes}
\usage{
dagri_blocked(graph)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}
}
\description{
Returns a named list of blocked nodes mapped to their block reasons. Call
\code{dagri_recompute_state()} before using this function to ensure node
states are current.
}
