% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{dagri_eligible}
\alias{dagri_eligible}
\title{Get eligible nodes}
\usage{
dagri_eligible(graph)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}
}
\description{
Returns IDs of nodes whose state is "ready". Call \code{dagri_recompute_state()}
before using this function to ensure node states are current.
}
