% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{dagri_pending_gates}
\alias{dagri_pending_gates}
\title{Get pending gates}
\usage{
dagri_pending_gates(graph, targets = NULL)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{targets}{Optional target nodes.}
}
\description{
Get pending gates
}
