% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{dagri_target_closure}
\alias{dagri_target_closure}
\title{Get the structural closure of target nodes}
\usage{
dagri_target_closure(graph, targets = NULL)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{targets}{Optional target nodes.}
}
\description{
Get the structural closure of target nodes
}
