% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{dagri_terminal}
\alias{dagri_terminal}
\title{Get terminal nodes}
\usage{
dagri_terminal(graph, targets = NULL)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{targets}{Optional target nodes.}
}
\description{
Get terminal nodes
}
\details{
O(V+E).
}
